requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
torch>=2.1.0
openai-whisper>=20231117
faster-whisper>=0.9.0
websockets>=12.0
numpy>=1.24.0
